Feature: Clone configs/disks (#5182)
* M3-2247 Clone Configs/Disks: Part 1 (Routing) (#5046)

* Routing for CloneLanding component

This covers the initial router for these components. <Configs /> and <Disks /> are placeholders for now.

* Wrap <Config /> and <Disk /> in <Paper />

* Change routing so CloneLanding is within Context

I realized that CloneLanding should be within the LinodeDetailContext so we have access to configs, disks, and the Linode itself.

My guess is that we'll eventually want to create a "Clone" tab or similar in LinodeDetail anyway.

* Fix imports

* Remove unused imports

* M3-2247 Part 2: Components and internal state (#5051)

* Add superCompact mode to Button

* Clone <Details /> component

* Clone <Disks /> component

* Clone <Configs /> component

* Update CloneLanding component

* Show LinodeSelect in Details

* Styles and useReducer pattern

* Clean

* Add empty states

* Change mapping so disks keep track of configs

* Clean

* Fix LinodeSelect null bug

* Fix unclear comment

* adding styling for longer labels and close alignment

* M3 2247 Part 3: Functionality (#5058)

* Add ability to exclude Linodes from LinodeSelect

* Add clone linode disk API method

* Add functionality, error/loading states

* Clean comments

* Clear error on input (review feedback)

* M3-2247 Part 2.5 (#5064)

* Expanded header for Linode Clone

This also allows custom props to be passed in to <LinodeControls />. I thought about using a `renderBreadcrumb` prop, but it's handy to be able to use the edit handlers built in to <LinodeControls />.

* Change order of Config and Disk action menu

* M3-2247 LinodeSelect group by region (#5070)

* Add `groupByRegion` prop to LinodeSelect

* Use  in Clone Details

* M3-2247 Review feedback (#5114)

* Change breadcrumb and add Clone title

* Update copy

* Custom error states

These errors don't always include a `field`, but we want to treat them as field errors, so some custom logic is required. Also include includes to resize your Linode and disks.

* Add estimated clone time

* Add tests and comments

* M3-2247: Unit Tests for Clone Configs/Disks (#5148)

* Change order of tests to match file

* Add unit tests for cloneLandingReducer

* Type diskConfigMap correctly

This shouldn't actually affect anything, but it's at least cleaner.

* Add tests for other utility functions

* Tests for Configs.tsx

* Add Disks.tsx tests

* Allow noCap prop on crumbOverrides

Previously the `noCap` prop worked only on the label (given via `labelTitle`). This allows specific crumbs (via crumbOverrides) to use noCap.

* Update MUI

This feature branch was created before the big MUI update was merged (#5008). This commit fixes all MUI-related errors after I rebased with the MUI update.

* Allow parent of LinodeSelect to pass value

This was misunderstood and subsequently removed in the merge commit. This adds the logic back in.

* Fix MD formatting

* Prop handling for DefaultLoader

Honestly... I'm not totally sure why this works. I was getting TS errors when passing props to Loadable component (in a `render` function of a <Route />). This is what ended up working.:

* Add code-splitting

[Review feedback from @martinmckenna]
Adds code-splitting for CloneLanding, LinodesDetailHeader, LinodesDetailNavigation, Configs, and Disks.

* Add disk_duplicate event messages

[Review feedback from @martinmckenna]

* Clean up a few comments

* Clean a few more comments

* Handle Clone events

* Remove import and add notification event handling for linode_clone