Get all SSH keys from /users endpoint
Original logic used in the SSH key HOC was first hitting /profile/ssh-keys
to get the current user's keys, then hitting /users to get keys for the
remaining users. Since the string label of each key was the only information
being retrieved, this was duplicative (and required an extra step to manually
skip the current user's results when looping through the list returned by
/users). It also led to inconsistent behavior, since /profile/ssh-keys is a
paginated list but the SSH keys returned form /users include the full list.

We have eliminated this logic, which simplifies the logic and makes the
returned values consistent.