M3-2879: Update Redux state shape for Domains (#5197)
* Use RequestableDataWithEntityError for Domains

* Add test fix

* Use filtered data (review feedback)

Instead of filtering twice.

* Use API response for result num (review feedback)

This required changing the signature of the getDomainsSuccess action creator to accept the number of results.

* Correctly count results in deleteDomain reducer

The `results` count was incorrect. Only filter data once.

[Note] This deleteDomain action is only used in domains.events. It can probably be removed (#tdt).

* Use `results` in DomainsDashboardCard

[review feedback]

* Fix DashboardCard error

Update the DashboardCard to use the new error shape (error.read). Type `withUpdatingDomains` correctly to guard against this sort of thing.