M3-3044 Change: Make search link the first option in Algolia search bar
In an attempt to make the search landing more prominent, move the
final search option in the Algolia search bar on the Get Help landing
page, which is a link to the search landing page, to the top of the
results list.

Now, when users type in a search query and press Enter, they'll be
taken to the search landing page. If they click on one of the results
in the dropdown, or scroll down to an option with the arrow keys,
they'll be taken (as before) to the external link for that result.