Convert BucketDetail to Class component
This was to fix a bug where the previous scroll position would be restored when navigating back from a folder. This re-render was happening before the data actually changed, which resulted in a jarring UX. I'm not totally sure why this hook magic is behaving the way it is, but converting to a class component fixes it, because when the component unmounts and remounts, the state is reset.