M3-3157 Fix: Config memory size limit not displayed
A line of code passing the memory_limit to its text field had been commented out.
As a result, opening a config for editing didn't display the value, leading
users to think that their edits had not been saved.

No idea why this line was commented out in the first place; must have been a reason
but I hope it's no longer valid.