Use overflow:auto instead of scroll
This fixed an issue where I was always seeing the scrollbar, even if it wasn't needed. Oddly enough, Alban was not seeing this, using the same browser. Turns out that it was a system issue, and the fact that I was using a USB mouse instead of a trackpad. For that reason, OS X was always showing the scrollbar for me. Using overflow:auto instead of overflow:scroll solved this issue.