Bucket detail enhancements (#5475)
* Bucket detail enhancements

* Add folder icon

* Convert BucketDetail to Class component

This was to fix a bug where the previous scroll position would be restored when navigating back from a folder. This re-render was happening before the data actually changed, which resulted in a jarring UX. I'm not totally sure why this hook magic is behaving the way it is, but converting to a class component fixes it, because when the component unmounts and remounts, the state is reset.

* Truncate utilities

* BucketBreadcrumb component

* Fix failing test

* Clean

* Clean comment

* switch icon

* add copied logic

* improve table styles

* Use overflow:auto instead of scroll

This fixed an issue where I was always seeing the scrollbar, even if it wasn't needed. Oddly enough, Alban was not seeing this, using the same browser. Turns out that it was a system issue, and the fact that I was using a USB mouse instead of a trackpad. For that reason, OS X was always showing the scrollbar for me. Using overflow:auto instead of overflow:scroll solved this issue.

* Make util function args generic

* Add comment explaining maxLength

* Copy bucketName + prefix to clipboard

As opposed to copying JUST the prefix. This was a quick fix to a problem where we showed the copy icon at the top-level, even though there was no prefix to copy.

However, this works just as well.