Feature/Bucket Detail (#5457)
* OBJ-58 Part I: Routes and directory structure (#5397)

* Move around structure

* Move BucketTableRow to the correct place

* Use rowLink prop instead of component

* Re-add redirect

* Move data-fetching back to landing

* Remove unnecessary ghost component

* Changes for better diff

* More changes for better diff

* Directory renaming

* OBJ-58 Part II: Bucket Detail page (#5403)

* Initial pass at Object table

* clean

* Add object icon

* Remove pagination

* OBJ-58 Part III: Folder Traversal (#5440)

* Pagination and traversal

* Refactor for clarity

* Remove unused function and comment

* Add comment

* Add tests and comments

* Review feedback

- Use constant for delimiter
- Change copy for empty folder state
- Clear data before requesting new data due to prefix change

* Use APIError from SDK

* Bucket detail enhancements (#5475)

* Bucket detail enhancements

* Add folder icon

* Convert BucketDetail to Class component

This was to fix a bug where the previous scroll position would be restored when navigating back from a folder. This re-render was happening before the data actually changed, which resulted in a jarring UX. I'm not totally sure why this hook magic is behaving the way it is, but converting to a class component fixes it, because when the component unmounts and remounts, the state is reset.

* Truncate utilities

* BucketBreadcrumb component

* Fix failing test

* Clean

* Clean comment

* switch icon

* add copied logic

* improve table styles

* Use overflow:auto instead of scroll

This fixed an issue where I was always seeing the scrollbar, even if it wasn't needed. Oddly enough, Alban was not seeing this, using the same browser. Turns out that it was a system issue, and the fact that I was using a USB mouse instead of a trackpad. For that reason, OS X was always showing the scrollbar for me. Using overflow:auto instead of overflow:scroll solved this issue.

* Make util function args generic

* Add comment explaining maxLength

* Copy bucketName + prefix to clipboard

As opposed to copying JUST the prefix. This was a quick fix to a problem where we showed the copy icon at the top-level, even though there was no prefix to copy.

However, this works just as well.

* Sort objects after getting them from the API

* Clean timer on unmount

* Throttle useWindowDimensions by default