Add Object Storage to PAT scopes
This should be feature flagged OFF until we are safely in OBJ GA.

This required changing the APIToken utility functions to take in `perms` as an argument, since the list of scopes is no longer static. In my estimation, we should keep this pattern around even after OBJ GA, since we will likely need it for other kinds of scopes in the future.