Fix messaging bug in Events rows
EventRow and ActivityRow both append the user who triggered an event
to the message generated by eventMessageGenerator(). This works in almost
all cases. However, there is a bug caused by the one exception:

`lassie_reboot` events already have "by the Linode watchdog service" as part
of the generated message (because this information is intended to be shown in the
bell events menu). This works fine there, but in the event rows the text is
confusing and repetitive, i.e. "Linode has been rebooted by the Linode
watchdog service by Linode".

Added a check for the event action to prevent this. To prevent repetition,
extracted this logic to a separate function in Event.helpers.ts. Another
option would be to remove the Lassie messaging from eventMessageGenerator,
and change `event.username` to `Lassie watchdog service` if
`event.action === 'lassie_reboot'`.

Threw in some unit tests and typing updates as well.