Preload device map on Linode Rebuild tab
Classic Manager preloads all of your available disks to your
available device slots on the Rebuild menu, until you run out
of either disks or slots. This PR adapts this behavior to Cloud.

A new function is added, getDefaultDeviceMapAndCounter, which
uses the disks available on a Linode (from the LinodeDetailContextProvider)
to generate a disk map and counter to preload the component state.
This same function is used on success to reset the form.