M3-3011 show booted config (#5656)
* Fix messaging bug in Events rows

EventRow and ActivityRow both append the user who triggered an event
to the message generated by eventMessageGenerator(). This works in almost
all cases. However, there is a bug caused by the one exception:

`lassie_reboot` events already have "by the Linode watchdog service" as part
of the generated message (because this information is intended to be shown in the
bell events menu). This works fine there, but in the event rows the text is
confusing and repetitive, i.e. "Linode has been rebooted by the Linode
watchdog service by Linode".

Added a check for the event action to prevent this. To prevent repetition,
extracted this logic to a separate function in Event.helpers.ts. Another
option would be to remove the Lassie messaging from eventMessageGenerator,
and change `event.username` to `Lassie watchdog service` if
`event.action === 'lassie_reboot'`.

Threw in some unit tests and typing updates as well.

* Add secondary entity label to linode_boot and linode_reboot event messages

* Change "booting" to "is being booted"

This makes the text in EventRow and ActivityRow make more sense, e.g.
"Linode my-linode: config-profile is being booted by prod-test-001"

* Fix typing for secondary entities (should be Entity | null)

* explicitly add with config X to the event message for linode_boot and linode_reboot

* add missing space