M3 Enables pasting for glish (#5627)
* enables pasting for glish

Glish being remote access to the virtualised VGA and PS2 Mouse and PS2
Keyboard does not have a pasting functionality. This PR enables pasting
by listening for a paste event on the document. If the mouse is on the
canvas noVNC will capture the paste event and pasting will not proceed,
but if the mouse is on the document part of the window then pasting will
occur in the window.

This works by taking the text from the paste buffer, and "manually"
actuating the keys through noVNC in an automated fashion. Currently this
works on text only and anything that is supported by a standard 104 key
keyboard.

In addition this only works in one direction browser to Linode. The
other way is not possible with this implementation.

This PR used
https://gist.github.com/byjg/a6378edb420a1c654c5f27bb494ca1c8 as
a reference.

* respond to PR feedback and move functionality into a function

which will then be called when the event fires