Hotfix: Safe access event.secondary_entity
I assumed that all events would come back with a secondary_entity,
but events created prior to this change were not updated. This caused
a lot of noise in Sentry (millions of reported exceptions) since we
weren't safe accessing the field.

Added a helper method, safeSecondaryEntityLabel, which handles this.

We should revisit this logic, since it gets pretty repetitive.