Add handling for Managed customers
- Re-add a React.useCallback that had been lost in a bad merge conflict resolution

- Add special handling for Managed customers who run into their LV client limit.

Based on:

There is a quirk in how APIv4 handles Managed customers and their Longview subscriptions. According to Mike Saah:

...it seems like the thing that ultimately makes API inconsistent w/ itself is that we don’t actually set customer.longview_limit when upgrading a customer to managed
when you upgrade to managed, API will cancel existing longview subscriptions, and tell you that your plan is longview-100
it will also disallow you from changing your plan, so long as you’re managed. but, it doesn’t actually update customer.longview_limit, which is what it validates on when creating new longview clients

This leads to edge cases in Longview.

For example:

1. I am not a Managed customer. I sign up for LongviewPro 3 pack
2. I decide to sign up for Managed.
3. In Cloud, we are told that my client_limit is 100. If I try to create a 4th client,
however, it will fail.

The same thing is true when cancelling a Managed subscription: If I have 3 clients, then subscribe/cancel Managed, the API will tell me that I have a client_limit of 10, but I still can't create more than 3 clients.

For now, in Cloud we are working around this by displaying custom text for Managed users, and for other users using their current number of clients (when they receive a "too many clients" error) to display their plan limit. For example, if I have 3 clients and the API blocks me when I try to create a 4th, we show a modal saying "Your current plan allows you to have 3 clients," simply reporting the number of clients I already have.

This also involved abstracting the "Contact Support" message from LongviewPlans and using it for Managed customers in the SubscriptionDialog.