Hack fix for componentWillReceiveProps
The form set was being reset every time the component re-rendered, which was every time we got new events.

I tried several workarounds, but this was the most straightforward. It should be noted that using componentWillReceiveProps in this way is still considered an anti-pattern, and we should refactor all uses. I chose not to do anything too fancy here, though, since the refactoring of these components will likely require larger discussion.