Revert cell loading state for RDNS (#5777)
We realized that having moving the loading state from the entire table to the individual RDNS meant that, if you didn't have RDNS enabled, you'd see a loading spinner for a few seconds, then it would disappear, leaving nothing. Alban and I decided that the TableRowLoading component was ultimately a better choice here.