Hide Longview tabs
Several tabs for Longview will not be included in the MVP release,
so this PR hids them under a separate LD flag (so that we can activate
longview without showing these tabs). Using a flag allows us to continue
to access the tabs in development without showing them to the user.

- Added a <Redirect /> to LongviewDetail to match behavior elsewhere.