M3-3624 Move events polling interval etc. out of Redux
- I previously moved setRequestDeadline, setPollingInterval and related
logic out of events.ts, because this had resulted in a circular dependency
(events imports the store which imports some of the methods above). Unfortunately,
this resulted in a bug where anything that subscribed to the events state in Redux
would render constantly as the polling interval was updated.

To try to avoid both issues, I created a new file at src/eventsPolling
that includes getters and setters for the variables in question.

- Resolved a circular dependency where ExtendedVolume was being imported
from child components.

- Fixed a typo in GroupedVolumes