Fix graphs padding
Longview graph data does not include forward padding for time intervals
longer than the available data range (if you ask for a start time
of 7 days ago but your Linode is new and only has 20 minutes of data,
you just get the 20 minutes). This causes our charts components to
set the X axis as 20 minutes. We can avoid this by adding a [startTime, null]
point to the beginning of the series.

This introduces some complications, though, when for example you're asking
for 30 minutes of data but the first point is 27 minutes ago. This should
be a smooth chart (there's probably a data point from ~32 minutes ago with
similar data) but our trick will add a gap to the start of the chart. To handle
this, we're checking if the difference between the indicated start time and
the first data point returned from the LV API is > 5 minutes. This may
not work for Longview Pro clients, although in my testing the resolution
is still 5 minutes.

Additionally, moved the data formatting function in OverviewGraphs
to a React.useCallback so that we can avoid calling it too often.