Longview e2e tests
Longview UI tests

* Created test for the longview plan subscription and longview clients.
* All tests can be run individually or together. The beforeEach and API calls will set up the data for the tests to run and use the information
from the responses for verification.

**note** there is an ugly browser.pause(1500) in the changePlan function.
This is because once the button is selected it is set to a disabled mode
so the check for the button to not be enabled will not work. Need a better
way to wait for the loading state on the button.

When running tests remember to use a user that does NOT have longview
clients installed. The tests will clear out all longview clients to make
sure the tests run reliably.

Added in some needed selectors to allow for UI automation to work easier. We are now using the data-testid instead of data-qa for testing attribiutes.

Fixes for crumb/header changes
changed title selector from '[data-qa-dialog-title]' to 'h1'