PR Fixes
Added a data-qa-loading selector that captures true/false on the loading state of the button. This allowed us to now wait for the button status to not be loading and remove the browser.pause(1500) line.
Added a second plan change selection to verify the loading wait and all is working.
Thanks to Pat for finding issues with some typo's for selectors (missing an open or close bracket) in the beforeEach that were causing the question mark status on the first longview plan test.

There is no longer a question mark on the first longview plan test and now three tests will fail when using a managed user for the test run. The failures will display the message: "cannot use a managed account for basic longview plan changes: Expected true to be false."