M3-3719 longview notices (#5932)
* Update types and send full LV response to consumers

* Refactor Longview to use full response

We were previously using handleLongviewResponse to separate data
from errors, and pass only API field errors (if there were any
NOTIFICATIONS) or the response DATA to consuming components.

The drawback was that many notifications are not errors, and
there was no way in our existing pipeline to handle these.

To address this, I passed the full, unaltered LongviewResponse
to any consumers. This required a large overhaul of our typings,
but consuming components seem to retain access to their pre-existing
type completion. Any actual errors (defined as NOTIFICATION.SEVERITY of 3)
are rejected as before, but as full NOTIFICATION objects rather than
APIError[]. Since we never used the API errors, this should not affect
consumers.

Since notifications are client-specific (afaict), I added a handler
for them in useClientLastUpdated, and used this value from within
LongviewDetail to populate a Notice for each notification.

* Type fixes

* Rename method (fix typo)

* Update pathOr to look for TEXT instead of reason to reflect new error shape.