Add copy explaining prorated transfer
This was causing confusion for some customers. The total amount of transfer shown on the Dashboard is prorated for the month. For example, if you had a 1GB Linode halfway during the month, 500GB is added to your transfer instead of 1000GB.

To clear up confusion, we include the note from Classic that explains that monthly transfer is prorated.