Add security.txt file (#5953)
Copy the existing security.txt file used for Linode.com and make it
available at the standard path (/.well-known/security.txt). Guidelines
for a repo say that it should be at the root level, and I couldn't figure
out how to do this with a symlink, so I just pasted it there too. It's a
very stable file so having multiple copies shouldn't be an issue.