Add empty state for Disks tab in LongviewDetail
Add a Placeholder component with empty state messaging when a user
has no disks that are reporting to Longview. This is different from
most empty states in LV, since most other tabs will have empty tables
or graphs. Disks, however, is an object of `sda: {data}` objects,
which will be empty if no data is available for the specified time frame.

We had discussed hiding the time select component from view in this empty
state, but on further testing discovered that it's easy to end up in a situation
where the empty state will disappear if the user changes to a longer time
frame, one in which data is present.