M3-3770 Convert invoice fields to Number() before using toFixed()
invoiceItem.amount (and .tax and .total) are guaranteed by the API
to be numbers, but some special cases (Customer Packages) and older
records returned these values as strings, which causes toFixed()
to throw an error.

The API will fix this bug in ARB-1607, but for safety sake we
should continue to manually convert values.

If anything other than a number or a valid numeric string comes in,
we'll end up showing "$NaN" to the user. I don't think this is possible
so I didn't do any complex conditional logic to account for it, but
we should keep this in mind in case something like this happens again.