Fix polling logic in useGraphs
The loading state was being set to true, and data being cleared,
every time the request() function was run. This conflicted with the
addition of polling, since checking for new data (which we're doing
to clear an error state if we get good data back) was triggering a
loading state every 10 seconds.

Added an optional isLoading param to request that defaults to true.
When false, as is the case in the polling useEffect() logic,
loading will not be set to true, and the data will not be reset.

This way, choosing a new time box shows a loading state and refreshes
the data, as intended, but the polling happens in the background.