M3 lv disks update logic (#5970)
* Remove unused imports

* Update ref typing in LineGraph

* Update Line mock in setupTests.js

* Fix refresh data logic in useGraphs

- Added an interval to the useGraphs component, since if e.g. requests
were blocked/unblocked, it wouldn't register this. (most of the consumers
were running this in a useEffect that listened for start, end, lastUpdated,
error, etc, but that wasn't penetrating).

- Add mounted checking to useGraphs to match useClientLastUpdated

- Removed feature flag from Disks tab, since if this PR is merged
we're good to release that feature.

* Factor isToday variable to a utility

* Remove shouldDelay prop to avoid odd UI when toggling the TimeRangeSelect

* Fix polling logic in useGraphs

The loading state was being set to true, and data being cleared,
every time the request() function was run. This conflicted with the
addition of polling, since checking for new data (which we're doing
to clear an error state if we get good data back) was triggering a
loading state every 10 seconds.

Added an optional isLoading param to request that defaults to true.
When false, as is the case in the polling useEffect() logic,
loading will not be set to true, and the data will not be reset.

This way, choosing a new time box shows a loading state and refreshes
the data, as intended, but the polling happens in the background.

* Add suggestedMax of disk size for Storage graphs

* Add correct suggested max to iNodes

* Add correct suggested max to iNodes

* Review feedback
- Update/add comments
- Replace pathOr with new TS magic