M3-3779 Allow editing IP transfer addresses from Slave edit form
We had a field to allow users to update axfr IPs (addresses allowed
to transfer for this Domain) when editing the SOA record for a master
Domain, but nothing was available for slave domains, since there's
only a single drawer for all slave-related editing actions.

Abstracted the axfr textfield so that it could be used in both places.
Also updated the warning text to more closely match Classic (there
is a limit of 6 IPs that we didn't mention).