M3-3465 Allow rebooting from Rescue mode
The API logic that determines which config to reboot with
when no config ID is provided seems to glitch out when
in Rescue mode. If a user has a single config, we just call
reboot(), which returns an API error "No bootable config found"
API is aware of this now, but since it's not going to get fixed
in time for EOL, the simplest thing for us to do is pass the configId
manually whenever only one config is present.

- I also renamed "Submit" to "Reboot into Rescue Mode" on the rescue button
to match Classic.