Use asyncWithLDProvider
The goal here was to be able to specify the API resource page_size with Launch Darkly feature flags.

The main issue with this was that we were making our initial requests BEFORE we request flags from LD. So a few things here:

- In order to use flags from the AuthenticationWrapper component, I had to lift the FeatureFlagProvider higher (it’s now at the root level).
- I also used the new `asyncWithLDProvider` which blocks rendering the rest of the app until the flags are fetched (this is what we want).
- The above required upgrading launchdarkly-react-client-sdk to v2.17.0
- This meant our patch couldn’t be applied, but I don’t think we need a patch anymore anyway.

There’s one wrinkle with this: we are now making a request to LD before we are redirected to Login. If this is unacceptable we’ll need to reconfigure the structure of AuthenticationWrapper.