M3-2446 Region status message
Add a banner to the top of all relevant pages (and to the support
ticket drawer) if any region returns a status of "outage". If multiple
regions have outages, the notices will stack.

This was our first use of the region.status field, so we had to update
typings and test data to respect the new API response.