M3-3465 Allow rebooting from Rescue mode (#6001)
* M3-3465 Allow rebooting from Rescue mode

The API logic that determines which config to reboot with
when no config ID is provided seems to glitch out when
in Rescue mode. If a user has a single config, we just call
reboot(), which returns an API error "No bootable config found"
API is aware of this now, but since it's not going to get fixed
in time for EOL, the simplest thing for us to do is pass the configId
manually whenever only one config is present.

- I also renamed "Submit" to "Reboot into Rescue Mode" on the rescue button
to match Classic.

* Add unit tests

* Refactor selectDefaultConfig

* Add configId to requests from LinodeCard view

In order to do this, added state management to LinodeCard. There's
now a loading state that is triggered when you click the Reboot
button, while we request config data for the selected Linode.

If the request succeeds, we open the reboot modal and proceed
using the returned config data. If it fails, we toast and encourage
the user to request again.

* Add comment to explain the configs request in LinodeCard