Classify restores from backups as rebuild events
When a Linode is restored from a backup, the IDs of its
configs and disks change. When the process is complete,
we need to re-request both configs and disks to ensure
that our cache is up to date. If we fail to do this, a
user will try to boot the Linode with a previous config,
which will cause an error.

Fortunately, we already have an event that has this behavior
(for the same reasons), linode_rebuild. By moving backup_restore
to the same case in the switch in linode.events.ts, we can
achieve the desired effect.