React.lazy (#5976)
* React.lazy/suspense POC in Kubernetes section

* Do things the easy way

* Add React.lazy to the Volumes index page

* Add abstracted loading component with delay

Create a new component, SuspenseLoader, that renders a CircleProgress
spinner after a delay. Delay is customizable via prop.

Wrapped the root-level <Route> component in MainContent with React.Suspense,
using the new loading component as a fallback. This allows us to
remove any calls to React.Suspense below in the tree (I think).

* Add delay to dependency array