Fix routing for slave domains
Previously clicking on a Slave Domain from the Search Landing page took you to the Domain Detail page for that Domain. The Details page isn't functional for Slave Domains, however. Here is one possible solution: simply navigating to the Domain Landing page when clicking a Slave Domain (as is the case with Volumes).