Add onSuccess handler to DeleteDomain
I realized that the onSuccess actions should be context-sensitive. For example, from the Search Results page, clicking the "Delete Domain" button in the drawer shouldn't navigate you to the Domain's page. In this commit I've introduced an optional `onSuccess` prop where consumers can do what they'd like. Note that the Toast is created by the DeleteDomain component, meaning the toast will *always* be created, regardless of what the `onSuccess` handler looks like. I figured this was desirable, but that's just my opinion.