Make StackScriptBase respect and update qs
Now when you select a StackScript tab ("Account" or "Community") the query string is updated with your selection. This means specific tabs can be bookmarked and pressing the browser's "Back" button will take you to the previous tab.

Additionally, when a user types in the Search field, the query string will be updated, which means searches can be bookmarked. The history is replaced, meaning pressing the browser's "Back" button will not take into account these events.