M3-3800 fix: Slave Domain routing from Search Landing (#6028)
* Fix routing for slave domains

Previously clicking on a Slave Domain from the Search Landing page took you to the Domain Detail page for that Domain. The Details page isn't functional for Slave Domains, however. Here is one possible solution: simply navigating to the Domain Landing page when clicking a Slave Domain (as is the case with Volumes).

* Make the drawer open instead

* Move ConditionalWrapper to /components

* Revert changes to ResultRow

Part of review feedback to handle Slave Domains with the routing.

* Allow domian container to omit mapState fn

* Handle routing for Slave Domains at the root lvl

* Singular instead of plural

* Adjust comment

* Fix imports

* Pass routerProps to domain detail and landing

Co-authored-by: Jared <jaredkobos@gmail.com>