Handle Lish sessions opened manually
If a user bookmarks a Lish url and opens a session that way,
window.close() won't work (and in fact will pollute the console
with error logs).

To handle this case, wrap window.close() in a try/catch block
and render an ErrorState if there is no token in localStorage.
This called for an "authenticated" value in state, which we
adjust in the setInterval.