Fix regression with Notistack
We recently upgraded to Notistack 0.9.7, and our logic update was incorrect.
This introduced a regression where clicking on the X on a toast message
dismissed all notices currently on the screen.

To fix this, I used patterns from the official docs
(https://iamhosseindhv.com/notistack/demos#action-for-all-snackbars).
This involved changing our Snackbar abstraction to a function component
and passing an onClick handler down to CloseSnackbar (rather than the key prop
that was previously used).

It may be possible to use the useSnackbar hook with this new version,
but I felt safer using officially recommended methods.