Fix SVG bug
There was a bug in which collapsing the primary nav caused a few of the heavenly-bucket.svg PATHs to have the incorrect fill. I honestly don't know what was happening here, but I think it had something to do with the link to the #a linear gradient, which is where the colors lived. To fix this, I just copied those colors to the linearGradients that referenced #a.