Refactor firewalls.reducer
- Refactor firewalls.reducer to use immer pattern
- Remove the notion of a "sequence" for a firewall along with the
associated types, because the API no longer cares about the order
in which rules are applied.
- Make listOfIdsInOriginalOrder from EntitiesAsObjectState, and remove
all use of it in our reducers. This was put in mainly for Firewalls
(see above), and was cluttering up every other reducer that used this
pattern. If we end up needing it for some reason, we can re-add it.