M3-3467 Fix updating Disks bug (#6073)
* Fix updating Disks bug

This fixes an alarming bug in which editing disks for one Linode in one tab modifies the *display* of disks for a different Linode in a different tab.

The mechanism behind this faulty behavior is that we have ONE key for Linode Disks in the store, called `__resources.linodeDisks`. This state is populated when a user navigates to a particular Linode’s Detail page. By design, only disks belonging to _that particular Linode_ are loaded in this state.

However, we dispatch the `getAllLinodeDisks` action when we receive a Disk event, so if we get a `disk_update` event for LinodeA, the `linodeDisks` state gets populated with LinodeA’s disks, even if we’re currently viewing LinodeB.

To fix, I have changed the shape of `linodeDisks`. Instead of being an instance of MappedEntityState, we map Linode IDs to their disks:

{
  [linodeA’s ID]: <Disks for LinodeA>,
  [linodeB’s ID]: <Disks for LinodeB>,
}

This required reworking the reducer, which Immer.js made clearer.

* Add unit tests

* Refactor for clarity and add comments

* Adjust loading and error

* Review feedback