M3-3942 Fix: Mode persistence in Linode Disk drawer
We use the same Disk drawer for all cases, disabling/hiding unneeded fields
for actions such as renaming a disk, where only one field is editable.

However, we were not resetting the "mode" of the drawer on close. This lead
to a bug:

1. Open the "Add Disk" drawer and choose "from Image"
2. Close the drawer
3. Reopen the drawer in "rename" or "resize" mode.
4. The "from image" fields are all there and editable.

This PR resets the mode to "from empty" on close.

- Fixes a small typo