M3-3857 firewall sdk redux (#6108)
* Refactor firewalls.reducer

- Refactor firewalls.reducer to use immer pattern
- Remove the notion of a "sequence" for a firewall along with the
associated types, because the API no longer cares about the order
in which rules are applied.
- Make listOfIdsInOriginalOrder from EntitiesAsObjectState, and remove
all use of it in our reducers. This was put in mainly for Firewalls
(see above), and was cluttering up every other reducer that used this
pattern. If we end up needing it for some reason, we can re-add it.

* Add partial factories and tests

* Fix more typings and the firewalls.reducer.tests

* Add logic and tests for deleting and updating firewall actions

* Fix things broken in rebase

* Add Redux request for Firewalls up one level to firewalls/index.tsx

* Review feedback 1

- Remove duplicate reducer case
- Adjust test case for clarity
- Replace oddball typing in FirewallTableRows with an explicit interface

* Fix rebase syntax snafu

* Remove listOfIDsInOriginalOrder

* Change 2 to 3 in a test case