M3-3966 PromotionalOffer component (#6102)
* Add PromotionalOffer interface and factory

* Add assets

* Add PromotionalOffer component

* Add PromotionalOffer on Dashboard

* Add PromotionalOffer on OBJ Landing

* Review feedback

* Add indicator to Primary Nav

* Conditionally display promos

* No nav icon in compact mode

* Remove icon from primary nav

[Review feedback]

* Fix SVG bug

There was a bug in which collapsing the primary nav caused a few of the heavenly-bucket.svg PATHs to have the incorrect fill. I honestly don't know what was happening here, but I think it had something to do with the link to the #a linear gradient, which is where the colors lived. To fix this, I just copied those colors to the linearGradients that referenced #a.

* Button support

* use feature flag instead of factory

* Remove None feature type since we use arrays now

* Add tests

* Max width

* URL Handing

Now, internal links use the React Router Link component so the app doesn't reload.
External links are opened in a new tab.