M3-3685 Display invalid SS errors on Linode rebuild
We noticed that selecting an invalid stackscript when rebuilding didn't
result in an error that the user can see. It turns out what was wrong
was that the field returned in that case was script, whereas our
Formik field was listening for stackscript_id.

I also tweaked the styles of SelectStackscriptPanel so that the
resulting error wasn't stuck to the top of the Paper.