M3-3409 LKE Node Pool Error
Show an error if node pools fail to load from the API. Currently
we show an empty state in the NodePoolsDisplayTable. This is likely
a consequence of that component having been originally designed for
the creation flow, where no node pool data from the API is needed.

We already store error, loading, and lastUpdated for node pools in
Redux. I refactored the display table to use a separate component,
NodePoolTableContent, to manage loading/error/empty/data states.
I then passed in the error and lastUpdated data using useSelector()
and displayed a table error state iff we have an api error from the node
pools endpoint and lastUpdated is 0 (so we've neer successfully loaded
this data).