Update syntax highlighting
Our Markdown parser and related logic are all fine and up to date, but the extension we've been
using for code highlighting hasn't been updated in 3 years. It also has a large footprint in our bundle.

To address this, I added highlight.js, which was powering the extension, as a direct dependency
and removed the extension completely. We can code split on highlight.js to only support the languages
we need. I checked with Support and code snippets in anything other than Bash are rare, so I kept
the language count small to maintain a smaller bundle (hopefully).

The logic we were using to sanitize and inject HTML was the same in all places, so I created a single
component, HighlightedMarkdown, to manage this process. The only drawback is normal text will be rendered
as HTML, which the support ticket reply display had logic to circumvent. The output doesn't appear different
so I figured it was a good trade.

- Add highlight.js as a direct dependency

- Add highlightedmarkdown component

- update HighlightedMarkdown to use Showdown parser

- Add support for yaml highlighting

- Use HighlightedMarkdown for support ticket reply formatting

- Add useMemo for markdown parsing

- Use HighlightedMarkdown for support ticket reply formatting phase 2

- Add Jest mock for highlight.js

- Add support for additional languages

- Add optional language prop to HighlightedMarkdown to specify language