Review feedback
- Delay clearing out the select state (by resetting the `key`)
until we actually receive an error. This improves the UX, since
otherwise the Promise.map could take a long time, leaving users
looking at an empty select and wondering what happened.

John suggested putting it in a .catch(), but the .catching() was already
being done in the parent, so I decided it was easiest to make clearing
the select a side-effect of receiving a new error message.