Source results from paginated request
Refactored volumes.reducer to use MappedEntityState2, our planned
type for all Redux state. This allows us to reference results
from the Dashboard instead of relying on data.length, which
is inaccurate now that we're only requesting a single page of data.

To support this, added a new version of store.helpers, store.helpers.tmp,
which contains modified versions of the functions in store.helpers, updated
to handle the new data shape.