Fix snapshot tests
RTL's asFragment() method depends on document.createRange().createDocumentFragment()
and we need to us asFragment() to run a snapshot test. However, due to a separate
MUI issue we were overriding document.createRange in testSetup.ts. I moved the override
to the test file that needed it so that snapshotting would work as expected.